requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
PyJWT==2.10.1
Pillow==11.0.0
Werkzeug==3.1.3
schedule==1.2.2
google-api-python-client==2.178.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
bcrypt==4.2.1

